algorithm: start to learn backtracking
